# swagger.yaml
openapi: 3.0.0
info:
  title: Crypto Price and Alert Service
  description: An API for managing cryptocurrency prices, alerts, and notifications.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
paths:
...
  /swap-rate:
    get:
      summary: Get ETH to BTC swap rate
      operationId: getSwapRate
      parameters:
        - name: ethAmount
          in: query
          required: true
          schema:
            type: number
          description: The amount of Ethereum to swap
      responses:
        200:
          description: Swap rate information
          content:
            application/json:
              schema:
                type: object
                properties:
                  btcAmount:
                    type: number
                    description: Amount of Bitcoin received for the swap
                  feeInEth:
                    type: number
                    description: Fee charged in Ethereum
                  feeInUsd:
                    type: number
                    description: Fee charged in USD